backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
python-dotenv==1.0.0
email-validator==2.1.0
numpy==1.24.3
redis==7.0.1
